pixels-to-image/requirements.txt
json5==0.9.6
Pillow==9.1.0